Move FunctionSignature to display/function_signature
